refactor: break down profile lookup into modular structure
- Split large vertex.ts file (1015 lines) into focused modules
- Create profile/ directory with specialized modules:
  - cache.ts: DVM and NIP-05 caching utilities
  - utils.ts: Profile event creation and field extraction
  - nip05.ts: NIP-05 resolution and verification
  - dvm-core.ts: Core DVM query functionality
  - dvm-lookup.ts: DVM lookup with fallback logic
  - fallback.ts: NIP-50 profile search fallback
  - resolver.ts: Unified author resolution
  - search.ts: Full-text profile search with ranking
  - index.ts: Centralized exports
- Maintain backward compatibility via vertex.ts re-exports
- All files under 210 lines as required
- Keep code DRY with shared utilities
- Improve maintainability and testability
