+++ src/lib/profile/resolver.ts
@@ -1,0 +1,62 @@
+import { NDKEvent } from '@nostr-dev-kit/ndk';
+import { nip19 } from 'nostr-tools';
+import { resolveNip05ToPubkey } from './nip05';
+import { profileEventFromPubkey } from './utils';
+import { lookupVertexProfile } from './dvm-lookup';
+import { fallbackLookupProfile } from './fallback';
+
+// Unified author resolver: npub | nip05 | username -> pubkey (hex) and an optional profile event
+export async function resolveAuthor(authorInput: string): Promise<{ pubkeyHex: string | null; profileEvent: NDKEvent | null }> {
+  try {
+    const input = (authorInput || '').trim();
+    if (!input) return { pubkeyHex: null, profileEvent: null };
+
+    // 1) If input is npub, decode directly
+    if (/^npub1[0-9a-z]+$/i.test(input)) {
+      try {
+        const { type, data } = nip19.decode(input);
+        if (type === 'npub' && typeof data === 'string') {
+          return { pubkeyHex: data, profileEvent: await profileEventFromPubkey(data) };
+        }
+      } catch {}
+      return { pubkeyHex: null, profileEvent: null };
+    }
+
+    // 2) If input looks like NIP-05 ('@name@domain' | 'domain.tld' | '@domain.tld'), resolve to pubkey
+    const nip05Like = input.match(/^@?([^\s@]+@[^\s@]+|[^\s@]+\.[^\s@]+)$/);
+    if (nip05Like) {
+      const pk = await resolveNip05ToPubkey(input);
+      if (!pk) return { pubkeyHex: null, profileEvent: null };
+      return { pubkeyHex: pk, profileEvent: await profileEventFromPubkey(pk) };
+    }
+
+    // 3) Otherwise treat as username and try Vertex DVM with fallback (single DVM attempt)
+    let profileEvt: NDKEvent | null = null;
+    try {
+      profileEvt = await lookupVertexProfile(`p:${input}`, fallbackLookupProfile);
+    } catch {}
+    if (!profileEvt) {
+      return { pubkeyHex: null, profileEvent: null };
+    }
+    const pubkeyHex = profileEvt.author?.pubkey || profileEvt.pubkey || null;
+    return { pubkeyHex, profileEvent: profileEvt };
+  } catch {
+    return { pubkeyHex: null, profileEvent: null };
+  }
+}
+
+// Resolve a by:<author> token value (username, nip05, or npub) to an npub.
+// Returns the original input if it's already an npub, otherwise attempts Vertex DVM
+// and falls back to a NIP-50 profile search. Hard timebox externally when needed.
+export async function resolveAuthorToNpub(author: string): Promise<string | null> {
+  try {
+    const input = (author || '').trim();
+    if (!input) return null;
+    if (/^npub1[0-9a-z]+$/i.test(input)) return input;
+    const { pubkeyHex } = await resolveAuthor(input);
+    if (!pubkeyHex) return null;
+    try { return nip19.npubEncode(pubkeyHex); } catch { return null; }
+  } catch {
+    return null;
+  }
+}
