--- src/lib/search/nip50.ts
@@ -1,81 +1,0 @@
-import { Nip50Extensions } from './types';
-
-// Extract NIP-50 extensions from the raw query string
-export function extractNip50Extensions(rawQuery: string): { cleaned: string; extensions: Nip50Extensions } {
-  let cleaned = rawQuery;
-  const extensions: Nip50Extensions = {};
-
-  // include:spam - turn off spam filtering
-  const includeSpamRegex = /(?:^|\s)include:spam(?:\s|$)/gi;
-  if (includeSpamRegex.test(cleaned)) {
-    extensions.includeSpam = true;
-    cleaned = cleaned.replace(includeSpamRegex, ' ');
-  }
-
-  // domain:<domain> - include only events from users whose valid nip05 domain matches the domain
-  const domainRegex = /(?:^|\s)domain:([^\s]+)(?:\s|$)/gi;
-  cleaned = cleaned.replace(domainRegex, (_, domain: string) => {
-    const value = (domain || '').trim();
-    if (value) extensions.domain = value;
-    return ' ';
-  });
-
-  // language:<two letter ISO 639-1 language code> - include only events of a specified language
-  const languageRegex = /(?:^|\s)language:([a-z]{2})(?:\s|$)/gi;
-  cleaned = cleaned.replace(languageRegex, (_, lang: string) => {
-    const value = (lang || '').trim().toLowerCase();
-    if (value && value.length === 2) extensions.language = value;
-    return ' ';
-  });
-
-  // sentiment:<negative/neutral/positive> - include only events of a specific sentiment
-  const sentimentRegex = /(?:^|\s)sentiment:(negative|neutral|positive)(?:\s|$)/gi;
-  cleaned = cleaned.replace(sentimentRegex, (_, sentiment: string) => {
-    const value = (sentiment || '').trim().toLowerCase();
-    if (['negative', 'neutral', 'positive'].includes(value)) {
-      extensions.sentiment = value as 'negative' | 'neutral' | 'positive';
-    }
-    return ' ';
-  });
-
-  // nsfw:<true/false> - include or exclude nsfw events (default: true)
-  const nsfwRegex = /(?:^|\s)nsfw:(true|false)(?:\s|$)/gi;
-  cleaned = cleaned.replace(nsfwRegex, (_, nsfw: string) => {
-    const value = (nsfw || '').trim().toLowerCase();
-    if (value === 'true') extensions.nsfw = true;
-    else if (value === 'false') extensions.nsfw = false;
-    return ' ';
-  });
-
-  return { cleaned: cleaned.trim(), extensions };
-}
-
-// Build search query with NIP-50 extensions
-export function buildSearchQueryWithExtensions(baseQuery: string, extensions: Nip50Extensions): string {
-  if (!baseQuery.trim()) return baseQuery;
-  
-  let searchQuery = baseQuery;
-  
-  // Add NIP-50 extensions as key:value pairs
-  if (extensions.includeSpam) {
-    searchQuery += ' include:spam';
-  }
-  
-  if (extensions.domain) {
-    searchQuery += ` domain:${extensions.domain}`;
-  }
-  
-  if (extensions.language) {
-    searchQuery += ` language:${extensions.language}`;
-  }
-  
-  if (extensions.sentiment) {
-    searchQuery += ` sentiment:${extensions.sentiment}`;
-  }
-  
-  if (extensions.nsfw !== undefined) {
-    searchQuery += ` nsfw:${extensions.nsfw}`;
-  }
-  
-  return searchQuery;
-}
