--- src/lib/search/types.ts
@@ -1,40 +1,0 @@
-import { NDKEvent, NDKFilter } from '@nostr-dev-kit/ndk';
-
-// Type definitions for relay objects
-export interface RelayObject {
-  url?: string;
-  relay?: {
-    url?: string;
-  };
-}
-
-// NIP-50 extension options
-export interface Nip50Extensions {
-  includeSpam?: boolean;
-  domain?: string;
-  language?: string;
-  sentiment?: 'negative' | 'neutral' | 'positive';
-  nsfw?: boolean;
-}
-
-export interface NDKEventWithRelaySource extends NDKEvent {
-  relaySource?: string;
-  relaySources?: string[]; // Track all relays where this event was found
-}
-
-// Extend filter type to include tag queries for "t" (hashtags)
-export type TagTFilter = NDKFilter & { '#t'?: string[] };
-
-// Streaming search options
-export interface StreamingSearchOptions {
-  exact?: boolean;
-  streaming?: boolean;
-  maxResults?: number;
-  timeoutMs?: number;
-  onResults?: (results: NDKEvent[], isComplete: boolean) => void;
-}
-
-// Centralized media extension lists (keep DRY)
-export const IMAGE_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif', 'gifs', 'apng', 'webp', 'avif', 'svg'] as const;
-export const VIDEO_EXTENSIONS = ['mp4', 'webm', 'ogg', 'ogv', 'mov', 'm4v'] as const;
-export const GIF_EXTENSIONS = ['gif', 'gifs', 'apng'] as const;
